[feat #175] code refactor
