+++ app/client/src/views/inspector/Dashboard.tsx
@@ -19,26 +19,15 @@ import { TableRow } from "@/components/table/TableRow";
 import { DropdownMenu } from "@/components/DropdownMenu/DropdownMenu";
 import { SmallNumberCard } from "@/components/Card/SmallNumberCard";
 
+import { formatDate } from "@/utils/dates";
+
+
+import type { LandData } from "@/types/interfaces";
+
+
 type ValuePiece = Date | null;
 type Value = [ValuePiece, ValuePiece];
 
-interface LandsCount {
-  total: number, 
-  registered: number,
-  bought:number,
-  unapproved:number
-} 
-
-function formatDate(date:Date){
-  const monthNames = [
-    "Jan", "Feb", "Mar", "Apr", "May", "Jun",
-    "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"
-  ];
-  const monthIndex = date.getMonth()
-  const dayIndex = date.getDate()
-
-  return `${monthNames[monthIndex]} ${dayIndex}`
-}
 
 function formatTimestampToDate(timestamp:number) {
   const date = new Date(timestamp * 1000); // Convert seconds to milliseconds
@@ -53,13 +42,7 @@ function formatTimestampToDate(timestamp:number) {
 }
 
 
-interface LandData {
-  landId?: string, 
-  area: number|null,
-  landUse: string,
-  latitude: number|null,
-  longitude: number|null
-}
+
 
 
 export function DashboardInspectorView() {
